requirements.txt
docling[vlm]==2.55.1
mistralai>=0.4.0
python-dotenv>=1.0.0
pytest>=8.0.0
pypdf>=4.0.0
openai>=1.0.0
langchain-core>=0.1.0
PyYAML>=6.0.1
tiktoken>=0.7.0
